# BUG/DEV str2vec to handle spaceless strings, performance - #305