--- a/serpentTools/tests/test_utils.py
+++ b/serpentTools/tests/test_utils.py
@@ -60,6 +60,13 @@ class VectorConverterTester(TestCase):
         expected = [0, 1, 2, 3]
         self._runConversionTest(int, expected, list)
 
+    def test_vecOfStr(self):
+        """Verify a single word can be converted with str2vec"""
+        key = 'ADF'
+        expected = array('ADF')
+        actual = str2vec(key)
+        assert_array_equal(expected, actual)
+
     def _runConversionTest(self, valType, expected, outType=None):
         if outType is None:
             outType = array
@@ -67,7 +74,7 @@ class VectorConverterTester(TestCase):
         else:
             compareType = outType
         for case in self.testCases:
-            actual = str2vec(case, of=valType, out=outType)
+            actual = str2vec(case, dtype=valType, out=outType)
             self.assertIsInstance(actual, compareType, msg=case)
             ofRightType = [isinstance(xx, valType) for xx in actual]
             self.assertTrue(all(ofRightType),