Closes #305 by allowing the conversion to handle string without spaces and create arrays of strings. When we perform `iterable.split()` on the string `'ADF'`, we get back a string that cannot be converted to floats or ints. This PR would instead return a direct array (or list or tuple) of the string.

## Performance

Utilize the python `map` function and `numpy.fromiter` to improve performance of the conversion. The following simple bash script compares two ways to construct a list of floats, and two ways to construct and array of floats. All commands work with the same range of values. 
```
#!/bin/bash
N=${1-10000}
echo Testing with $N values

python -m timeit -s "xr = range(${N})" "list(map(float, xr))"
python -m timeit -s "xr = range(${N})" "[float(xx) for xx in xr]"
python -m timeit -s "from numpy import fromiter; xr = range(${N})" "fromiter(map(float, xr), float)"
python -m timeit -s "from numpy import array; xr=range(${N})" "array([float(xx) for xx in xr])"
```
The times are ordered as:
1. List of floats using map
2. List of floats using list comprehension
3. Array of floats using map
4. Array of floats using list comprehension
```
Testing with 10 values
500000 loops, best of 5: 751 nsec per loop
500000 loops, best of 5: 913 nsec per loop
200000 loops, best of 5: 1.23 usec per loop
200000 loops, best of 5: 1.71 usec per loop
Testing with 100 values
50000 loops, best of 5: 5.3 usec per loop
50000 loops, best of 5: 7.54 usec per loop
50000 loops, best of 5: 6.9 usec per loop
20000 loops, best of 5: 10.8 usec per loop
Testing with 1000 values
5000 loops, best of 5: 59.5 usec per loop
5000 loops, best of 5: 83.3 usec per loop
5000 loops, best of 5: 69.6 usec per loop
2000 loops, best of 5: 111 usec per loop
Testing with 10000 values
500 loops, best of 5: 605 usec per loop
500 loops, best of 5: 837 usec per loop
500 loops, best of 5: 692 usec per loop
200 loops, best of 5: 1.08 msec per loop
```
We can see that, for small values, the two are comparable. As the number of entries increases the `map`-based approach becomes comparatively faster. This speed up is nice when people read universe data with 100s of energy groups, or detectors with thousands of groups. This function is called a lot, and this marginal speed up *should* make a difference in the scheme of reading. We don't have a profiling and timing suite yet, but we should.